TwittMap_2.1_Kafka/requirements.txt
boto3==1.4.4
Flask==0.11.1
kafka_python==1.3.3
elasticsearch==5.2.0
tweepy==3.5.0
watson_developer_cloud==0.25.2
googlemaps==2.4.6